Bug 970199 - Switch mouse event button type from ushort to short according to spec change. r=smaug
UltraBlame original commit: ac8adf61a96b73ad57d0b30fdebab5dc86154fc9

+++ editor/libeditor/base/nsEditorEventListener.cpp
@@ -587,7 +587,7 @@ nsEditorEventListener::MouseClick(nsIDOMEvent* aMouseEvent)
   
   mEditor->ForceCompositionEnd();
 
-  uint16_t button = (uint16_t)-1;
+  int16_t button = -1;
   mouseEvent->GetButton(&button);
   
   if (button == 1)
